requirements.txt
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
django-heroku==0.3.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
psycopg2==2.7.4
PyJWT==1.6.0
pytz==2018.3
whitenoise==3.3.1
